% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deploy.R
\name{deploy}
\alias{deploy}
\title{Deploy app within inst directory to shinyapps.io via rsconnect}
\usage{
deploy(file, title, ...)
}
\arguments{
\item{file}{Folder location for app (e.g., "/inst/apps/ascent-operations")}

\item{title}{Name of the app (will be part of the url)}

\item{...}{Further parameters passed to rsconnect::deployApp (e.g., account)}
}
\value{
Invisibly returns `NULL`; called for its side effect of deploying
  the app to shinyapps.io
}
\description{
Deploy app within inst directory to shinyapps.io via rsconnect
}
