% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_db_connection}
\alias{get_db_connection}
\title{Connect to the ASCENT Postgres database}
\usage{
get_db_connection(config_name, ...)
}
\arguments{
\item{config_name}{Name of the `config.yml` configuration block holding
the database `dbname`, `user`, `password`, `host`, and `port`}

\item{...}{Further arguments passed to `config::get()`}
}
\value{
A `DBIConnection` object connected to the ASCENT Postgres
  database
}
\description{
Connect to the ASCENT Postgres database
}
