% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{get_flux_client}
\alias{get_flux_client}
\title{Connect to the ASCENT InfluxDB (Flux) instance}
\usage{
get_flux_client(config_name)
}
\arguments{
\item{config_name}{Name of the `config.yml` configuration block holding
the `influx_read_token`}
}
\value{
An `InfluxDBClient` connected to the ASCENT InfluxDB instance
}
\description{
Connect to the ASCENT InfluxDB (Flux) instance
}
