% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map_ascent_static}
\alias{map_ascent_static}
\title{A little bespoke map for plotting values at all of the ascent sites. To use this,
export at desired pixel dimensions, then use a tool like MS Paint to determine the
pixel locations of the sites. These can be translated to coordinates in ggplot.}
\usage{
map_ascent_static(con)
}
\arguments{
\item{con}{A database connection, as returned by [get_db_connection()]}
}
\value{
A `ggplot`/`cowplot` object showing site locations
}
\description{
Draws a static map of the contiguous US with all ASCENT site locations
plotted, plus inset maps for Alaska (Delta Junction) and Southern
California (Los Angeles, Rubidoux, Joshua Tree).
}
