% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smps.R
\name{smps_data}
\alias{smps_data}
\title{Retrieve raw SMPS sample analysis data for a site and time range}
\usage{
smps_data(site, start_dt, end_dt, con)
}
\arguments{
\item{site}{ASCENT site code}

\item{start_dt}{Start date/datetime (inclusive) of the requested range}

\item{end_dt}{End date (inclusive) of the requested range}

\item{con}{A database connection, as returned by [get_db_connection()]}
}
\value{
A data frame of SMPS sample analysis records
}
\description{
Retrieve raw SMPS sample analysis data for a site and time range
}
