% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acsm.R
\name{acsm_autoqc_file}
\alias{acsm_autoqc_file}
\title{Apply auto-qc checks to ACSM L1a data and return a file for external merging with ACSM
HDF data in Igor. This process requires adjusting the time stamps to match.}
\usage{
acsm_autoqc_file(site, start_dt, end_dt, con)
}
\arguments{
\item{con}{}
}
\description{
Apply auto-qc checks to ACSM L1a data and return a file for external merging with ACSM
HDF data in Igor. This process requires adjusting the time stamps to match.
}
