% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{map_ascent_static}
\alias{map_ascent_static}
\title{A little bespoke map for plotting values at all of the ascent sites. To use this,
export at desired pixel dimensions, then use a tool like MS Paint to determine the
pixel locations of the sites. These can be translated to coordinates in ggplot.}
\usage{
map_ascent_static(con)
}
\arguments{
\item{con}{}
}
\description{
Title
}
