% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{monthly_l1b}
\alias{monthly_l1b}
\title{Title}
\usage{
monthly_l1b(site, month_date, instrument, folder, con, influx_con = NULL)
}
\arguments{
\item{site}{}

\item{month_date}{}

\item{instrument}{}

\item{folder}{}

\item{con}{}

\item{influx_con}{}
}
\description{
Title
}
