% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smps_l2.R
\name{smps_l2_from_files}
\alias{smps_l2_from_files}
\title{Title}
\usage{
smps_l2_from_files(l1b_file, manual_qc_file, start_datetime = NULL)
}
\arguments{
\item{l1b_file}{}

\item{manual_qc_file}{}

\item{start_datetime}{}
}
\description{
Title
}
