% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xact.R
\name{xact_metadata}
\alias{xact_metadata}
\title{Title}
\usage{
xact_metadata(
  site,
  start_dt,
  end_dt,
  level = "1a",
  con,
  metadata_fields = NULL
)
}
\arguments{
\item{metadata_fields}{}
}
\description{
Title
}
